+++ src/Stage.js
@@ -44,7 +44,7 @@ ThreeRTT.Stage.prototype = {
   },
 
   // Add object render pass
-  paint: function (object) {
+  paint: function (object, empty) {
 
     // Create root to hold all objects for this pass
     var root = new THREE.Object3D();
@@ -52,9 +52,13 @@ ThreeRTT.Stage.prototype = {
     root.visible = true;
 
     // Create a surface to render the last frame
-    var material = new ThreeRTT.FragmentMaterial(this, 'generic-fragment-texture');
-    var surface = this._surface(material);
-    root.add(surface);
+    if (!empty) {
+      var material = new ThreeRTT.FragmentMaterial(this, 'generic-fragment-texture');
+      var surface = this._surface(material);
+      root.add(surface);
+    }
+
+    // Add object
     root.add(object);
 
     // Add root to scene and insert into pass list